3d-interation: initial commit.
This component is meant to be attached to 3D elements in the scene that are interactable. Currently it creates visual feedback when hovering and switches to the camera position. This requires the text element / box / etc. to have an ID that is identical to the Game State that it wants to switch to. Could be refined still.