ui-text-popup: initial commit.
This adds the functionality to display a custom popup when clicking on a certain button. The popup is defined in a `<div>` container with a corresponding style. Each of these have a matching id that is passed along in the parameters of the `ui-text-popup="popupId: Product-1-Popup"`. Also the overlay id can be passed along this way.