main-camera: access camera to rotate towards zero.
Get a reference of the actual camera object3D (three.js object) and rotate it towards the inverse of whatever rotation it is experiencing in the first x frames of the simulation.

The background is that the gyroscope rotation is applied to the a-frame components non-predictably within the first few frames. In the very first frame there is no rotation, it only "kicks in" gradually. This approach tries to counter act this by applying the inverse of the world rotation (gyro) to the a-frame object (i.e. the camera).