package.json: initial commit.
This is a necessary file for building the website (using `npm install` and `npm start`). The latter command will actually call the "start" entry in "scripts" within the `package.json`, which in turn uses some `budo` command to launch the app. This currently also has some information on the repository. The `package-lock.json` is automatically generated when building and contains all the dependencies that all the libraries use that we want to import and utilize.