tilt back screen functionality relocated and disabled
The tilt back screen functionality was meant to solve the bug that occurs when a user tilts his phone. When a user tilts his phone to enter wide screen format, the image gets distorted and even stays distorted when the screen is tilted back. the tilt back screen functionality is a temporal solve for this bug by displaying a "please tilt back phone" alert and reload the page. We had to disable this functionality, as we now see that it also triggers page force reload when the desktop window format changes. The bug remains unfixed at this point